connect wallets and pricing plans across services

+++ packages/models-library/src/models_library/api_schemas_webserver/wallets.py
@@ -23,7 +23,7 @@ class WalletGet(OutputSchema):
 
 
 class WalletGetWithAvailableCredits(WalletGet):
-    available_credits: float
+    available_credits: Decimal
 
 
 class WalletGetPermissions(WalletGet):
